requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.16
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
mysqlclient==2.2.5
pillow==10.4.0
sqlparse==0.5.1
typing_extensions==4.12.2
uWSGI==2.0.28
